+++ src/form/shared.jl
@@ -374,23 +374,28 @@ function constraint_switch_voltage_on_off(pm::AbstractWModels, n::Int, i, f_bus,
 end
 
 
-
 ""
-function constraint_current_limit(pm::AbstractWModels, n::Int, f_idx, c_rating_a)
+function constraint_current_limit_from(pm::AbstractWModels, n::Int, f_idx, c_rating_a)
     l,i,j = f_idx
-    t_idx = (l,j,i)
 
     w_fr = var(pm, n, :w, i)
-    w_to = var(pm, n, :w, j)
 
     p_fr = var(pm, n, :p, f_idx)
     q_fr = var(pm, n, :q, f_idx)
     JuMP.@constraint(pm.model, p_fr^2 + q_fr^2 <= w_fr*c_rating_a^2)
+end
+
+""
+function constraint_current_limit_to(pm::AbstractWModels, n::Int, t_idx, c_rating_a)
+    l,j,i = t_idx
+
+    w_to = var(pm, n, :w, j)
 
     p_to = var(pm, n, :p, t_idx)
     q_to = var(pm, n, :q, t_idx)
     JuMP.@constraint(pm.model, p_to^2 + q_to^2 <= w_to*c_rating_a^2)
 end
+
 
 ""
 function constraint_storage_losses(pm::AbstractWConvexModels, n::Int, i, bus, r, x, p_loss, q_loss)